src/requirements.txt
six
puremagic
PyPDF2
Pillow
flask
flask-login
Flask-WTF
Flask-Cache
Flask-Script
flask-markdown
flask_migrate
Flask-SQLAlchemy
mdx-linkify==1.0
mdx-journal>=0.1.4
flask-bcrypt
python-dotenv
